Avoid code in __init__
CMK-16672
